removed dbStore. Now remains the tree logic, which is called by the state.
